Refactored whole business layer. Separated presentation layer from business logic layer (mappings occur in controllers now). Separated query and command classes from their handlers. Created a repository for Operations.

+++ Domain/Contact.cs
@@ -14,7 +14,7 @@ namespace Domain
         public DateTime DateAdded { get; set; }
         public string Notes { get; set; }
         public string Status { get; set; }
-        public Int16 SuccessfulDeals { get; set; }
+        public int SuccessfulDeals { get; set; }
         public string Source { get; set; }
         public Boolean Premium { get; set; }
         public virtual ICollection<Order> Orders { get; set; }
